requirements.txt
streamlit>=1.40.0
anthropic>=0.40.0
yfinance>=0.2.40
pandas>=2.0.0
numpy>=1.24.0
plotly>=5.18.0
requests>=2.31.0
python-dotenv>=1.0.0
pyyaml>=6.0.0
APScheduler>=3.10.0
